+++ lib/roar/decorator.rb
@@ -2,18 +2,10 @@ require 'roar/representer'
 require 'representable/decorator'
 
 class Roar::Decorator < Representable::Decorator
-  extend Roar::Representer::InheritableArray
-
   module HypermediaConsumer
-    def links_array=(*args)
-      super # TODO: this currently sets #links which is not obvious.
-      represented.links = links
-    end
-
-    # TODO: what is the deal with #links_array and #links?
-    def links=(*args)
-      super
-      represented.links = links
+    def links=(arr)
+      links = super
+      represented.instance_variable_set :@links, links
     end
   end
 end